refactor secrets manager (general fixes)
