+++ src/zenml/secrets_manager/__init__.py
@@ -22,6 +22,8 @@ from zenml.secrets_manager.local.local_secrets_manager import (
     LocalSecretsManager,
 )
 
+# TODO [MEDIUM]: Lazy load the AWSSecretsManager instead of importing it here.
+
 __all__ = [
     "BaseSecretsManager",
     "LocalSecretsManager",
