+++ src/zenml/secrets_manager/local/local_secrets_manager.py
@@ -15,6 +15,7 @@ import os
 from typing import Any, Dict, List, Optional
 
 from zenml.cli.utils import error
+from zenml.constants import LOCAL_SECRETS_FILENAME
 from zenml.enums import SecretsManagerFlavor, StackComponentType
 from zenml.io.fileio import create_file_if_not_exists
 from zenml.io.utils import get_global_config_directory
@@ -22,15 +23,14 @@ from zenml.logger import get_logger
 from zenml.secret import SecretSchemaClassRegistry
 from zenml.secret.base_secret import BaseSecretSchema
 from zenml.secrets_manager.base_secrets_manager import BaseSecretsManager
-from zenml.secrets_manager.utils import decode_secret_dict, encode_secret
+from zenml.utils.secrets_manager_utils import decode_secret_dict, encode_secret
 from zenml.stack.stack_component_class_registry import (
     register_stack_component_class,
 )
 from zenml.utils import yaml_utils
 
-logger = get_logger(__name__)
 
-LOCAL_SECRETS_FILENAME = "secrets.yaml"
+logger = get_logger(__name__)
 
 
 @register_stack_component_class(
@@ -97,7 +97,9 @@ class LocalSecretsManager(BaseSecretsManager):
             raise KeyError(f"Secret set `{secret_name}` does not exists.")
         secret_dict = secret_sets_store_items[secret_name]
 
-        decoded_secret_dict, zenml_schema_name = decode_secret_dict(secret_dict)
+        decoded_secret_dict, zenml_schema_name = decode_secret_dict(
+            secret_dict
+        )
         decoded_secret_dict["name"] = secret_name
 
         secret_schema = SecretSchemaClassRegistry.get_class(
